Rename pegen_interface.h to parser_interface.h
